fix: refactored interceptor to filters
