+++ OneTrustConsentFilter/lib/src/main/res/values/colors.xml
