--- OneTrustInterceptor/lib/build.gradle
@@ -1,53 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'org.jetbrains.kotlin.android'
-}
-
-android {
-    namespace 'com.rudderstack.android.adapter.onetrustinterceptor'
-    compileSdk 32
-
-    defaultConfig {
-        minSdk 21
-        targetSdk 32
-        versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
-}
-
-dependencies {
-
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.4'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-    compileOnly 'com.onetrust.cmp:native-sdk:202212.1.0.0'
-    compileOnly 'com.rudderstack.android.sdk:core:1.8.1-SNAPSHOT'
-
-    testImplementation "org.hamcrest:hamcrest:2.2"
-    testImplementation "org.mockito:mockito-core:3.11.2"
-    testImplementation "org.powermock:powermock-core:2.0.9"
-    testImplementation "org.powermock:powermock-module-junit4:2.0.9"
-    testImplementation "org.powermock:powermock-api-mockito2:2.0.9"
-    testImplementation 'com.onetrust.cmp:native-sdk:202212.1.0.0'
-    testImplementation 'com.rudderstack.android.sdk:core:1.8.1-SNAPSHOT'
-    testImplementation 'com.google.code.gson:gson:2.8.6'
-    testImplementation 'org.json:json:20210307'
-
-
-}
